Release: Sprint 4

Closes #98

See merge request informatik/qe/swess23/group4/g4t3!60

+++ src/main/java/at/qe/skeleton/repositories/AccessPointRepository.java
@@ -1,8 +1,8 @@
 package at.qe.skeleton.repositories;
 
-import at.qe.skeleton.model.AccessPoint;
+import at.qe.skeleton.models.AccessPoint;
 
 public interface AccessPointRepository extends AbstractRepository<AccessPoint, Long> {
 
-    AccessPoint findFirstById(Long id);
+    AccessPoint findFirstById(Integer id);
 }
