+++ src/main/java/at/qe/skeleton/repositories/SensorStationRepository.java
@@ -1,8 +1,8 @@
 package at.qe.skeleton.repositories;
 
-import at.qe.skeleton.model.SensorStation;
+import at.qe.skeleton.models.SensorStation;
 
 public interface SensorStationRepository extends AbstractRepository<SensorStation, Long>{
 
-    SensorStation findFirstById(Long id);
+    SensorStation findFirstById(Integer id);
 }
