+++ src/main/java/at/qe/skeleton/repositories/SensorValuesRepository.java
@@ -1,8 +1,8 @@
 package at.qe.skeleton.repositories;
 
-import at.qe.skeleton.model.SensorValues;
+import at.qe.skeleton.models.SensorValues;
 
 public interface SensorValuesRepository extends AbstractRepository<SensorValues, Long> {
 
-    SensorValues findFirstById(Long id);
+    SensorValues findFirstById(Integer id);
 }
